requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers
djangorestframework==3.14.0
drf-yasg==1.21.7
sqlparse==0.4.4
typing_extensions==4.6.3
psycopg2-binary==2.9.6
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
rarfile==4.0
Unidecode==1.3.6
dedupe==2.0.23
Faker==18.11.2
gunicorn==20.1.0
colorlog==6.7.0
lxml==4.9.3
html5lib==1.1
pypdfium2==4.18.0
PyPDF2==3.0.1  # se sigue necesitando
tqdm
